requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
alembic==1.16.5
pydantic-settings==2.10.1
email-validator==2.2.0
bcrypt==4.2.1
argon2-cffi==23.1.0
PyYAML==6.0.2
Jinja2==3.1.6
httpx==0.28.1
pymysql==1.1.1
psycopg[binary]==3.2.10
qrcode[pil]==7.4.2
reportlab==4.2.5
python-multipart==0.0.20
websocket-client==1.8.0
